+++ docs/conf.py
@@ -158,3 +158,7 @@ texinfo_documents = [
         "Data",
     ),
 ]
+
+
+# also include __init__ constructor in the docs
+autoclass_content = "both"